+++ src/komponenter/header/header-regular/mobil/hovedmeny/meny-dropdown/mobil-innhold/innloggetbruker/InnloggetBruker.less
