+++ src/komponenter/header/header-regular/mobil/hovedmeny/meny-dropdown/mobil-innhold/top-seksjon/Toppseksjon.less
