+++ src/komponenter/header/skiplinks/Skiplinks.tsx
@@ -3,7 +3,7 @@ import { Element } from 'nav-frontend-typografi';
 import Tekst from 'tekster/finn-tekst';
 import { mobilviewMax } from '../../../styling-mediaquery';
 import { matchMedia } from 'utils/match-media-polyfill';
-import { mobilHovedmenyKnappId } from '../header-regular/mobil/hovedmeny-mobil/HovedmenyMobil';
+import { mobilHovedmenyKnappId } from '../header-regular/mobil/hovedmeny/HovedmenyMobil';
 import { desktopHovedmenyKnappId } from '../header-regular/desktop/hovedmeny/HovedmenyDesktop';
 import { desktopSokKnappId } from '../header-regular/desktop/sok/SokDropdown';
 import './Skiplinks.less';
