+++ src/utils/keyboard-navigation/kb-navigation-setup.ts
@@ -10,7 +10,7 @@ import { desktopHovedmenyKnappId } from 'komponenter/header/header-regular/deskt
 import { desktopHeaderLogoId } from 'komponenter/header/header-regular/desktop/DesktopMenylinje';
 import { desktopSokKnappId } from 'komponenter/header/header-regular/desktop/sok/SokDropdown';
 import { desktopVarslerKnappId } from 'komponenter/header/header-regular/desktop/varsler-desktop/VarslerDropdown';
-import { desktopMinsideKnappId } from 'komponenter/header/header-regular/desktop/minside-meny/MinsideMenyDesktop';
+import { desktopMinsideKnappId } from 'komponenter/header/header-regular/desktop/minside-meny/MinsideMeny';
 import { Language } from 'store/reducers/language-duck';
 import { MenuValue } from '../meny-storage-utils';
 import { Status } from 'api/api';
